:truck: rename turbus files schema
